+++ src/app/store/reducers.ts
@@ -29,7 +29,7 @@ export interface RootState {
   readonly dimApi: DimApiState;
 }
 
-export type ThunkResult<R> = ThunkAction<Promise<R>, RootState, {}, AnyAction>;
+export type ThunkResult<R = void> = ThunkAction<Promise<R>, RootState, {}, AnyAction>;
 export type ThunkDispatchProp = {
   dispatch: ThunkDispatch<RootState, {}, AnyAction>;
 };