+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* External wish lists will be checked daily. Settings menu shows last fetched time.
+
 ## 5.74.0 <span className="changelog-date">(2020-03-15)</span>
 
 * Updated item metadata for Season of the Worthy! Mod slots, sources, ghost perks, etc.